+++ src/CleanRegex/Internal/Factory/Worker/OffsetsWorker.php
@@ -6,7 +6,6 @@ use TRegx\CleanRegex\Exception\NoSuchNthElementException;
 use TRegx\CleanRegex\Exception\SubjectNotMatchedException;
 use TRegx\CleanRegex\Internal\Factory\Optional\NotMatchedOptionalWorker;
 use TRegx\CleanRegex\Internal\Factory\Optional\OptionalWorker;
-use TRegx\CleanRegex\Internal\Factory\Optional\SubjectOptionalWorker;
 use TRegx\CleanRegex\Internal\Messages\NthOffsetMessage;
 use TRegx\CleanRegex\Internal\Messages\Subject\FirstMatchOffsetMessage;
 use TRegx\CleanRegex\Internal\Messages\Subject\NthMatchOffsetMessage;
@@ -46,7 +45,7 @@ class OffsetsWorker implements StreamWorker
 
     public function noNth(int $nth, int $total): OptionalWorker
     {
-        return new SubjectOptionalWorker(new NthOffsetMessage($nth, $total), $this->subject, NoSuchNthElementException::class);
+        return new NotMatchedOptionalWorker(new NthOffsetMessage($nth, $total), $this->subject, $this->notMatched, NoSuchNthElementException::class);
     }
 
     public function unmatchedNth(int $nth): OptionalWorker